Onboarding_Post/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-firestore
google-cloud-storage
firebase-admin
pillow
langchain
langchain-openai